+++ R/googleAuthR_shiny.R
@@ -12,6 +12,7 @@
 #'
 #' @return Shiny UI
 #' @export
+#' @importFrom shiny tagList tags HTML NS
 gar_auth_jsUI <- function(id, 
                           login_class = "btn btn-primary",
                           logout_class = "btn btn-danger",
@@ -20,14 +21,14 @@ gar_auth_jsUI <- function(id,
 
   
   check_package_loaded("shiny")
-  ns <- shiny::NS(id)
+  ns <- NS(id)
 
-  shiny::tagList(
+  tagList(
 
-    shiny::tags$script(src='https://apis.google.com/js/auth.js'),
-    shiny::tags$button(id = ns("login"), onclick="auth();", login_text, class = login_class),
-    shiny::tags$button(id = ns("logout"), onclick="out();", logout_text, class = logout_class),
-    shiny::tags$script(type="text/javascript", shiny::HTML(paste0("
+    tags$script(src='https://apis.google.com/js/auth.js'),
+    tags$button(id = ns("login"), onclick="auth();", login_text, class = login_class),
+    tags$button(id = ns("logout"), onclick="out();", logout_text, class = logout_class),
+    tags$script(type="text/javascript", HTML(paste0("
       var authorizeButton = document.getElementById('",ns("login"),"');
       var signoutButton = document.getElementById('",ns("logout"),"');
       signoutButton.style.display = 'none';
@@ -70,11 +71,12 @@ gar_auth_jsUI <- function(id,
 #'
 #' @return A httr reactive OAuth2.0 token
 #' @export
+#' @importFrom shiny validate need reactive req
 gar_auth_js <- function(input, output, session){
     check_package_loaded("shiny")
-    js_token <- shiny::reactive({
-      shiny::validate(
-        shiny::need(input$js_auth_access_token, "Authenticate")
+    js_token <- reactive({
+      validate(
+        need(input$js_auth_access_token, "Authenticate")
       )
       
       list(access_token = input$js_auth_access_token,
@@ -85,9 +87,9 @@ gar_auth_js <- function(input, output, session){
     })
     
     ## Create access token
-    access_token <- shiny::reactive({
+    access_token <- reactive({
       
-      shiny::req(js_token())
+      req(js_token())
       
       gar_js_getToken(js_token())
       
@@ -99,25 +101,27 @@ gar_auth_js <- function(input, output, session){
 
 #' Create a httr token from a js token
 #' @keywords internal
+#' @noRd
+#' @importFrom httr oauth_app Token2.0 oauth_endpoints
 gar_js_getToken <- function(token,
                             client.id     = getOption("googleAuthR.webapp.client_id"),
                             client.secret = getOption("googleAuthR.webapp.client_secret")){
   check_package_loaded("shiny")
-  gar_app <- httr::oauth_app("google", key = client.id, secret = client.secret)
+  gar_app <- oauth_app("google", key = client.id, secret = client.secret)
   
   scope_list <- getOption("googleAuthR.scope")
   
   # Create a Token2.0 object consistent with the token obtained from gar_auth()
   token_formatted <-
-    httr::Token2.0$new(app = gar_app,
-                       endpoint = httr::oauth_endpoints("google"),
-                       credentials = list(access_token = token$access_token,
-                                          token_type = token$token_type,
-                                          expires_in = token$expires_in,
-                                          refresh_token = NULL),
-                       params = list(scope = scope_list, type = NULL,
-                                     use_oob = FALSE, as_header = TRUE),
-                       cache_path = getOption("googleAuthR.httr_oauth_cache"))
+    Token2.0$new(app = gar_app,
+                 endpoint = oauth_endpoints("google"),
+                 credentials = list(access_token = token$access_token,
+                                    token_type = token$token_type,
+                                    expires_in = token$expires_in,
+                                    refresh_token = NULL),
+                 params = list(scope = scope_list, type = NULL,
+                               use_oob = FALSE, as_header = TRUE),
+                 cache_path = getOption("googleAuthR.httr_oauth_cache"))
   
   token_formatted
 }
@@ -146,10 +150,11 @@ createCode <- function(seed=NULL, num=20){
 #' @return The Google auth token in the code URL parameter.
 #' @family shiny auth functions
 #' @keywords internal
+#' @importFrom shiny parseQueryString
 authReturnCode <- function(session, 
                            securityCode=getOption("googleAuthR.securitycode")){
   check_package_loaded("shiny")
-  pars <- shiny::parseQueryString(session$clientData$url_search)
+  pars <- parseQueryString(session$clientData$url_search)
   
   if(!is.null(pars$state)){
     if(pars$state != securityCode){
@@ -182,6 +187,7 @@ authReturnCode <- function(session,
 #' 
 #' @family shiny auth functions
 #' @keywords internal
+#' @importFrom httr modify_url oauth_endpoints
 gar_shiny_getAuthUrl <- 
   function(redirect.uri,
            state = getOption("googleAuthR.securitycode"),
@@ -197,8 +203,8 @@ gar_shiny_getAuthUrl <-
     scopeEnc <- paste(scope, sep='', collapse=' ')
     
     ## httr friendly version
-    url <- httr::modify_url(
-      httr::oauth_endpoints("google")$authorize,
+    url <- modify_url(
+      oauth_endpoints("google")$authorize,
       query = list(response_type = "code",
                    client_id = client.id,
                    redirect_uri = redirect.uri,
@@ -255,354 +261,40 @@ gar_shiny_getUrl <- function(session){
 #' @keywords internal
 #' @return A list including the token needed for Google API requests.
 #' @family shiny auth functions
+#' @importFrom httr oauth_app POST headers content Token2.0 oauth_endpoints
 gar_shiny_getToken <- function(code,
                                redirect.uri,
                                client.id     = getOption("googleAuthR.webapp.client_id"),
                                client.secret = getOption("googleAuthR.webapp.client_secret")){
   
-  gar_app <- httr::oauth_app("google", key = client.id, secret = client.secret)
+  gar_app <- oauth_app("google", key = client.id, secret = client.secret)
   
   scope_list <- getOption("googleAuthR.scope")
   
   req <-
-    httr::POST("https://accounts.google.com/o/oauth2/token",
+    POST("https://accounts.google.com/o/oauth2/token",
                body = list(code = code,
                            client_id = client.id,
                            client_secret = client.secret,
                            redirect_uri = redirect.uri,
                            grant_type = "authorization_code"))
   
-  stopifnot(identical(httr::headers(req)$`content-type`,
+  stopifnot(identical(headers(req)$`content-type`,
                       "application/json; charset=utf-8"))
   # content of req will contain access_token, token_type, expires_in
-  token <- httr::content(req, type = "application/json")
+  token <- content(req, type = "application/json")
   
   # Create a Token2.0 object consistent with the token obtained from gar_auth()
-  token_formatted <-
-    httr::Token2.0$new(app = gar_app,
-                       endpoint = httr::oauth_endpoints("google"),
-                       credentials = list(access_token = token$access_token,
-                                          token_type = token$token_type,
-                                          expires_in = token$expires_in,
-                                          refresh_token = token$refresh_token),
-                       params = list(scope = scope_list, type = NULL,
-                                     use_oob = FALSE, as_header = TRUE),
-                       cache_path = getOption("googleAuthR.httr_oauth_cache"))
+  Token2.0$new(app = gar_app,
+               endpoint = oauth_endpoints("google"),
+               credentials = list(access_token = token$access_token,
+                                  token_type = token$token_type,
+                                  expires_in = token$expires_in,
+                                  refresh_token = token$refresh_token),
+               params = list(scope = scope_list, type = NULL,
+                             use_oob = FALSE, as_header = TRUE),
+               cache_path = getOption("googleAuthR.httr_oauth_cache"))
   
-  token_formatted
-}
-
-#' Create a reactive Google OAuth2 token
-#' 
-#' Use within a Shiny server.R session to create the access token passed 
-#'   to all Google API functions using \code{with_shiny}
-#' 
-#' @param session A Shiny session object.
-#' 
-#' @return A reactive Google auth token
-#' @family shiny auth functions
-#' @export
-#' @examples
-#' \dontrun{
-#' ## in global.R
-#' 
-#' ## create the API call function, example with goo.gl URL shortner
-#' library(googleAuthR)
-#' options("googleAuthR.scopes.selected" = c("https://www.googleapis.com/auth/urlshortener"))
-#' 
-#' shorten_url <- function(url){
-#' 
-#'   body = list(
-#'     longUrl = url
-#'  )
-#'  
-#'  f <- gar_api_generator("https://www.googleapis.com/urlshortener/v1/url",
-#'                         "POST",
-#'                         data_parse_function = function(x) x$id)
-#'                         
-#'  f(the_body = body)
-#'  
-#'  }
-#' 
-#' 
-#' ## in server.R
-#' library(shiny)
-#' library(googleAuthR)
-#' source('global.R')
-#' 
-#' shinyServer(function(input, output, session)){
-#'   
-#'   ## Get auth code from return URL
-#'   access_token  <- reactiveAccessToken(session)
-#' 
-#'   ## Make a loginButton to display using loginOutput
-#'   output$loginButton <- renderLogin(session, access_token())
-#'
-#'   short_url_output <- eventReactive(input$submit, {
-#'   ## wrap existing function with_shiny
-#'   ## pass the reactive token in shiny_access_token
-#'   ## pass other named arguments
-#'     short_url <- with_shiny(f = shorten_url, 
-#'                            shiny_access_token = access_token(),
-#'                            url=input$url)
-#'                            
-#'    })
-#'    
-#'    output$short_url <- renderText({
-#'    
-#'      short_url_output()
-#'      
-#'    })
-#'  }
-#' 
-#' ## in ui.R
-#' library(shiny)
-#' library(googleAuthR)
-#' 
-#' shinyUI(
-#'   fluidPage(
-#'     loginOutput("loginButton"),
-#'     textInput("url", "Enter URL"),
-#'     actionButton("submit", "Shorten URL"),
-#'     textOutput("short_url")
-#'     ))
-#' }
-reactiveAccessToken <- function(session){
-  check_package_loaded("shiny")
-  message("reactiveAccessToken is deprecated as of googleAuthR 0.3.0. Use googleAuth() and googleAuthUI() instead.")
-  shiny::reactive({
-    ## gets all the parameters in the URL. The auth code should be one of them.
-    
-    if(!is.null(authReturnCode(session))){
-      ## extract the authorization token
-      app_url <- gar_shiny_getUrl(session)    
-      access_token <- gar_shiny_getToken(authReturnCode(session), app_url)
-      
-      Authentication$set("public", "app_url", app_url, overwrite=TRUE)
-      Authentication$set("public", "shiny", TRUE, overwrite=TRUE)
-      
-      access_token
-      
-    } else {
-      NULL
-    }
-  })
-}
-
-#' Login/logout Shiny output
-#' 
-#' USe within a ui.R to render the login button generated by \code{renderLogin}
-#' 
-#' @param output_name Name of what output object was assigned in \code{renderLogin}
-#' 
-#' @return A login/logout button in a Shiny app
-#' 
-#' @export
-#' @family shiny auth functions
-#' @examples
-#' \dontrun{
-#' ## in global.R
-#' 
-#' ## create the API call function, example with goo.gl URL shortner
-#' library(googleAuthR)
-#' options("googleAuthR.scopes.selected" = c("https://www.googleapis.com/auth/urlshortener"))
-#' 
-#' shorten_url <- function(url){
-#' 
-#'   body = list(
-#'     longUrl = url
-#'  )
-#'  
-#'  f <- gar_api_generator("https://www.googleapis.com/urlshortener/v1/url",
-#'                         "POST",
-#'                         data_parse_function = function(x) x$id)
-#'                         
-#'  f(the_body = body)
-#'  
-#'  }
-#' 
-#' 
-#' ## in server.R
-#' library(shiny)
-#' library(googleAuthR)
-#' source('global.R')
-#' 
-#' shinyServer(function(input, output, session)){
-#'   
-#'   ## Get auth code from return URL
-#'   access_token  <- reactiveAccessToken(session)
-#' 
-#'   ## Make a loginButton to display using loginOutput
-#'   output$loginButton <- renderLogin(session, access_token())
-#'
-#'   short_url_output <- eventReactive(input$submit, {
-#'   ## wrap existing function with_shiny
-#'   ## pass the reactive token in shiny_access_token
-#'   ## pass other named arguments
-#'     short_url <- with_shiny(f = shorten_url, 
-#'                            shiny_access_token = access_token(),
-#'                            url=input$url)
-#'                            
-#'    })
-#'    
-#'    output$short_url <- renderText({
-#'    
-#'      short_url_output()
-#'      
-#'    })
-#'  }
-#' 
-#' ## in ui.R
-#' library(shiny)
-#' library(googleAuthR)
-#' 
-#' shinyUI(
-#'   fluidPage(
-#'     loginOutput("loginButton"),
-#'     textInput("url", "Enter URL"),
-#'     actionButton("submit", "Shorten URL"),
-#'     textOutput("short_url")
-#'     ))
-#' }
-loginOutput <- function(output_name){
-  check_package_loaded("shiny")
-  message("loginOutput is deprecated as of googleAuthR 0.3.0. Use googleAuth() and googleAuthUI() instead.")
-  shiny::uiOutput(output_name)
-}
-
-#' Render a Google API Authentication Login/logout button
-#' 
-#' Use within a Shiny server.R to assign to an output for ui.R.
-#' The login button carries an ActionLink with value "signed_in" 
-#'   but as Shiny reloads on pushing it can't be used for detection of login state.  
-#'   Use \code{!is.null(access_token())} instead.
-#' 
-#' @param session A Shiny session object
-#' @param access_token A token generated by \code{reactiveAccessToken}
-#' @param login_text What the login text will read on the button
-#' @param logout_text What the logout text will read on the button
-#' @param login_class The Bootstrap class for the login link
-#' @param logout_class The Bootstrap class for the logout link
-#' @param access_type Online or offline access for the authentication URL. 
-#' @param revoke If TRUE a user on logout will need to re-authenticate.
-#' @param approval_prompt Whether to show the consent screen on authentication.
-#' @return An object to assign to output e.g. output$login
-#' 
-#' @seealso \link{revokeEventObserver}
-#' 
-#' @export
-#' @family shiny auth functions
-#' @examples
-#' \dontrun{
-#' ## in global.R
-#' 
-#' ## create the API call function, example with goo.gl URL shortner
-#' library(googleAuthR)
-#' options("googleAuthR.scopes.selected" = c("https://www.googleapis.com/auth/urlshortener"))
-#' 
-#' shorten_url <- function(url){
-#' 
-#'   body = list(
-#'     longUrl = url
-#'  )
-#'  
-#'  f <- gar_api_generator("https://www.googleapis.com/urlshortener/v1/url",
-#'                         "POST",
-#'                         data_parse_function = function(x) x$id)
-#'                         
-#'  f(the_body = body)
-#'  
-#'  }
-#' 
-#' 
-#' ## in server.R
-#' library(shiny)
-#' library(googleAuthR)
-#' source('global.R')
-#' 
-#' shinyServer(function(input, output, session)){
-#'   
-#'   ## Get auth code from return URL
-#'   access_token  <- reactiveAccessToken(session)
-#' 
-#'   ## Make a loginButton to display using loginOutput
-#'   ## revoke=TRUE means upon logout a user will need to reauthenticate
-#'   output$loginButton <- renderLogin(session, access_token(), revoke=TRUE)
-#'   
-#'   ## Needed if revoke=TRUE above
-#'   revokeEventObserver(access_token())
-#'
-#'   short_url_output <- eventReactive(input$submit, {
-#'   ## wrap existing function with_shiny
-#'   ## pass the reactive token in shiny_access_token
-#'   ## pass other named arguments
-#'     short_url <- with_shiny(f = shorten_url, 
-#'                            shiny_access_token = access_token(),
-#'                            url=input$url)
-#'                            
-#'    })
-#'    
-#'    output$short_url <- renderText({
-#'    
-#'      short_url_output()
-#'      
-#'    })
-#'  }
-#' 
-#' ## in ui.R
-#' library(shiny)
-#' library(googleAuthR)
-#' 
-#' shinyUI(
-#'   fluidPage(
-#'     loginOutput("loginButton"),
-#'     textInput("url", "Enter URL"),
-#'     actionButton("submit", "Shorten URL"),
-#'     textOutput("short_url")
-#'     ))
-#' }
-renderLogin <- function(session, 
-                        access_token,
-                        login_text="Login via Google",
-                        logout_text="Logout",
-                        login_class="btn btn-primary",
-                        logout_class="btn btn-default",
-                        access_type = c("online","offline"),
-                        approval_prompt = c("auto","force"),
-                        revoke = FALSE){
-  check_package_loaded("shiny")
-  message("renderLogin is deprecated as of googleAuthR 0.3.0. Use googleAuth() and googleAuthUI() instead.")
-  access_type <- match.arg(access_type)
-  approval_prompt <- match.arg(approval_prompt)
-  
-  shiny::renderUI({
-    if(is.null(shiny::isolate(access_token))) {
-      shiny::actionLink("signed_in",
-                 shiny::a(login_text, 
-                          href = gar_shiny_getAuthUrl(gar_shiny_getUrl(session), 
-                                                      access_type = access_type,
-                                                      approval_prompt = approval_prompt), 
-                   class=login_class, 
-                   role="button"))
-    } else {
-      if(revoke){
-        
-        logout_button <- shiny::actionButton("revoke", "Revoke Access", 
-                                             href = gar_shiny_getUrl(session), 
-                                             class=logout_class,
-                                             role="button")
-
-      } else {
-        logout_button <- shiny::a(logout_text, 
-                                  href = gar_shiny_getUrl(session), 
-                                  class=logout_class, 
-                                  role="button")
-      }
-      
-      logout_button
-      
-    }
-  })
 }
 
 #' A Login button (Shiny Module)
@@ -615,11 +307,12 @@ renderLogin <- function(session,
 #' 
 #' @family shiny module functions
 #' @export
+#' @importFrom shiny NS uiOutput
 googleAuthUI <- function(id){
   check_package_loaded("shiny")
-  ns <- shiny::NS(id)
+  ns <- NS(id)
   
-  shiny::uiOutput(ns("googleAuthUi"))
+  uiOutput(ns("googleAuthUi"))
 }
   
   
@@ -700,6 +393,7 @@ googleAuthUI <- function(id){
 #' 
 #' @family shiny module functions
 #' @export
+#' @import shiny
 googleAuth <- function(input, output, session, 
                        login_text="Login via Google",
                        logout_text="Logout",
@@ -709,9 +403,10 @@ googleAuth <- function(input, output, session,
                        approval_prompt = c("auto","force"),
                        revoke = FALSE){
   check_package_loaded("shiny")
-  access_type <- match.arg(access_type)
+  
+  access_type     <- match.arg(access_type)
   approval_prompt <- match.arg(approval_prompt)
-  ns <- session$ns
+  ns              <- session$ns
   
   accessToken <- shiny::reactive({
     
@@ -777,101 +472,6 @@ googleAuth <- function(input, output, session,
 
 }
 
-#' Listens for a user revoking authentication
-#' 
-#' @description If the parameter \code{revoke} is set to TRUE for \link{renderLogin}
-#'   then this observer is also required in the Shiny server to do the revoking.
-#' 
-#' @param access_token A token generated by \code{reactiveAccessToken}.
-#' @param input the input object from a shinyServer function.
-#' 
-#' @seealso \link{renderLogin}
-#' 
-#' @export
-#' @family shiny auth functions
-#' @examples
-#' \dontrun{
-#' ## in global.R
-#' 
-#' ## create the API call function, example with goo.gl URL shortner
-#' library(googleAuthR)
-#' options("googleAuthR.scopes.selected" = c("https://www.googleapis.com/auth/urlshortener"))
-#' 
-#' shorten_url <- function(url){
-#' 
-#'   body = list(
-#'     longUrl = url
-#'  )
-#'  
-#'  f <- gar_api_generator("https://www.googleapis.com/urlshortener/v1/url",
-#'                         "POST",
-#'                         data_parse_function = function(x) x$id)
-#'                         
-#'  f(the_body = body)
-#'  
-#'  }
-#' 
-#' 
-#' ## in server.R
-#' library(shiny)
-#' library(googleAuthR)
-#' source('global.R')
-#' 
-#' shinyServer(function(input, output, session)){
-#'   
-#'   ## Get auth code from return URL
-#'   access_token  <- reactiveAccessToken(session)
-#' 
-#'   ## Make a loginButton to display using loginOutput
-#'   ## revoke=TRUE means upon logout a user will need to reauthenticate
-#'   output$loginButton <- renderLogin(session, access_token(), revoke=TRUE)
-#'   
-#'   ## Needed if revoke=TRUE above
-#'   revokeEventObserver(access_token())
-#'
-#'   short_url_output <- eventReactive(input$submit, {
-#'   ## wrap existing function with_shiny
-#'   ## pass the reactive token in shiny_access_token
-#'   ## pass other named arguments
-#'     short_url <- with_shiny(f = shorten_url, 
-#'                            shiny_access_token = access_token(),
-#'                            url=input$url)
-#'                            
-#'    })
-#'    
-#'    output$short_url <- renderText({
-#'    
-#'      short_url_output()
-#'      
-#'    })
-#'  }
-#' 
-#' ## in ui.R
-#' library(shiny)
-#' library(googleAuthR)
-#' 
-#' shinyUI(
-#'   fluidPage(
-#'     loginOutput("loginButton"),
-#'     textInput("url", "Enter URL"),
-#'     actionButton("submit", "Shorten URL"),
-#'     textOutput("short_url")
-#'     ))
-#' } 
-revokeEventObserver <- function(access_token, input){
-  check_package_loaded("shiny")
-  message("revokeEventObserver is deprecated as of googleAuthR 0.3.0. Use googleAuth() and googleAuthUI() instead.")  
-  shiny::observeEvent(input$revoke, {
-
-        ## GETS the revoke URL for this user's access_token
-        httr::GET(httr::modify_url("https://accounts.google.com/o/oauth2/revoke",
-                                   query = 
-                                   list(token = 
-                                        shiny::isolate(access_token)$credentials$access_token)))
-        myMessage("Revoked access", level=2)
-  })
-
-}
 
 #' Turn a googleAuthR data fetch function into a Shiny compatible one
 #' 
@@ -949,7 +549,8 @@ revokeEventObserver <- function(access_token, input){
 #' }
 with_shiny <- function(f, shiny_access_token=NULL, ...){
   if(is.null(shiny_access_token)) 
-    stop("Need to provide the reactive access token in shiny_access_token argument. e.g. shiny_access_token=access_token()")
+    stop("Need to provide the reactive access token in shiny_access_token argument. 
+         e.g. shiny_access_token=access_token()")
 
   formals(f) <- c(formals(f), list(shiny_access_token=shiny_access_token))